Fixed ForwardServer's echo mode by making the loopback sockets blocking after earlier pull request inadvertently changed them to non-blocking.
Implemented initial adapter_transport.PlainTransport acceptance tests.
